utils/cat_age_data.py
import torch
from torchvision import transforms
import pandas as pd
from PIL import Image
import os


class CatAgeDataset(torch.utils.data.Dataset):
    def __init__(self, csv_file, img_dir, transform=None):
        self.labels_df = pd.read_csv(csv_file)
        self.img_dir = img_dir
        self.transform = transform

    def __len__(self):
        return len(self.labels_df)

    def __getitem__(self, idx):
        img_name = os.path.join(self.img_dir, self.labels_df.iloc[idx, 0])
        image = Image.open(img_name)

        if self.transform:
            image = self.transform(image)

        age = self.labels_df.iloc[idx, 1]
        age = torch.tensor(float(age))

        return image, age
